Move all the pretty-print functions to their own file in cmd/notary.
Also add tests for pretty-printing and sorting targets.
