+++ cmd/notary/keys.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -17,7 +16,6 @@ import (
 	"github.com/docker/notary/trustmanager"
 
 	"github.com/docker/notary/tuf/data"
-	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 )
 
@@ -99,110 +97,6 @@ var cmdKeyRemove = &cobra.Command{
 	Short: "Removes the key with the given keyID.",
 	Long:  "Removes the key with the given keyID.  If the key is stored in more than one location, you will be asked which one to remove.",
 	Run:   keyRemove,
-}
-
-func truncateWithEllipsis(str string, maxWidth int, leftTruncate bool) string {
-	if len(str) <= maxWidth {
-		return str
-	}
-	if leftTruncate {
-		return fmt.Sprintf("...%s", str[len(str)-(maxWidth-3):])
-	}
-	return fmt.Sprintf("%s...", str[:maxWidth-3])
-}
-
-const (
-	maxGUNWidth = 25
-	maxLocWidth = 40
-)
-
-type keyInfo struct {
-	gun      string // assumption that this is "" if role is root
-	role     string
-	keyID    string
-	location string
-}
-
-// We want to sort by gun, then by role, then by keyID, then by location
-// In the case of a root role, then there is no GUN, and a root role comes
-// first.
-type keyInfoSorter []keyInfo
-
-func (k keyInfoSorter) Len() int      { return len(k) }
-func (k keyInfoSorter) Swap(i, j int) { k[i], k[j] = k[j], k[i] }
-func (k keyInfoSorter) Less(i, j int) bool {
-	// special-case role
-	if k[i].role != k[j].role {
-		if k[i].role == data.CanonicalRootRole {
-			return true
-		}
-		if k[j].role == data.CanonicalRootRole {
-			return false
-		}
-		// otherwise, neither of them are root, they're just different, so
-		// go with the traditional sort order.
-	}
-
-	// sort order is GUN, role, keyID, location.
-	orderedI := []string{k[i].gun, k[i].role, k[i].keyID, k[i].location}
-	orderedJ := []string{k[j].gun, k[j].role, k[j].keyID, k[j].location}
-
-	for x := 0; x < 4; x++ {
-		switch {
-		case orderedI[x] < orderedJ[x]:
-			return true
-		case orderedI[x] > orderedJ[x]:
-			return false
-		}
-		// continue on and evalulate the next item
-	}
-	// this shouldn't happen - that means two values are exactly equal
-	return false
-}
-
-// Given a list of KeyStores in order of listing preference, pretty-prints the
-// root keys and then the signing keys.
-func prettyPrintKeys(keyStores []trustmanager.KeyStore, writer io.Writer) {
-	var info []keyInfo
-
-	for _, store := range keyStores {
-		for keyPath, role := range store.ListKeys() {
-			gun := ""
-			if role != data.CanonicalRootRole {
-				gun = filepath.Dir(keyPath)
-			}
-			info = append(info, keyInfo{
-				role:     role,
-				location: store.Name(),
-				gun:      gun,
-				keyID:    filepath.Base(keyPath),
-			})
-		}
-	}
-	if len(info) == 0 {
-		writer.Write([]byte("No signing keys found.\n"))
-		return
-	}
-
-	sort.Stable(keyInfoSorter(info))
-
-	table := tablewriter.NewWriter(writer)
-	table.SetHeader([]string{"ROLE", "GUN", "KEY ID", "LOCATION"})
-	table.SetBorder(false)
-	table.SetColumnSeparator(" ")
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.SetCenterSeparator("-")
-	table.SetAutoWrapText(false)
-
-	for _, oneKeyInfo := range info {
-		table.Append([]string{
-			oneKeyInfo.role,
-			truncateWithEllipsis(oneKeyInfo.gun, maxGUNWidth, true),
-			oneKeyInfo.keyID,
-			truncateWithEllipsis(oneKeyInfo.location, maxLocWidth, true),
-		})
-	}
-	table.Render()
 }
 
 func keysList(cmd *cobra.Command, args []string) {
@@ -396,12 +290,6 @@ func keysImportRoot(cmd *cobra.Command, args []string) {
 	if err != nil {
 		fatalf("Error importing root key: %v", err)
 	}
-}
-
-func printKey(cmd *cobra.Command, keyPath, alias, loc string) {
-	keyID := filepath.Base(keyPath)
-	gun := filepath.Dir(keyPath)
-	cmd.Printf("%s - %s - %s - %s\n", gun, alias, keyID, loc)
 }
 
 func keysRotate(cmd *cobra.Command, args []string) {
