+++ cmd/notary/tuf.go
@@ -3,16 +3,13 @@ package main
 import (
 	"bufio"
 	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net"
 	"net/http"
 	"net/url"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"time"
 
@@ -25,7 +22,6 @@ import (
 	notaryclient "github.com/docker/notary/client"
 	"github.com/docker/notary/tuf/data"
 	"github.com/docker/notary/utils"
-	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 )
 
@@ -447,40 +443,4 @@ func getRemoteTrustServer() string {
 		}
 	}
 	return remoteTrustServer
-}
-
-type targetsSorter []*notaryclient.Target
-
-func (t targetsSorter) Len() int      { return len(t) }
-func (t targetsSorter) Swap(i, j int) { t[i], t[j] = t[j], t[i] }
-func (t targetsSorter) Less(i, j int) bool {
-	return t[i].Name < t[j].Name
-}
-
-// Given a list of KeyStores in order of listing preference, pretty-prints the
-// root keys and then the signing keys.
-func prettyPrintTargets(ts []*notaryclient.Target, writer io.Writer) {
-	if len(ts) == 0 {
-		writer.Write([]byte("\nNo targets present in this repository.\n\n"))
-		return
-	}
-
-	sort.Stable(targetsSorter(ts))
-
-	table := tablewriter.NewWriter(writer)
-	table.SetHeader([]string{"Name", "Digest", "Size (bytes)"})
-	table.SetBorder(false)
-	table.SetColumnSeparator(" ")
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.SetCenterSeparator("-")
-	table.SetAutoWrapText(false)
-
-	for _, t := range ts {
-		table.Append([]string{
-			t.Name,
-			hex.EncodeToString(t.Hashes["sha256"]),
-			fmt.Sprintf("%d", t.Length),
-		})
-	}
-	table.Render()
 }
